+++ src/com/timmattison/bitcoin/test/script/words/arithmetic/Op1Add.java
@@ -12,7 +12,7 @@ import com.timmattison.bitcoin.test.script.Word;
  */
 public class Op1Add extends Word {
     private static final String word = "OP_1ADD";
-    private static final int opcode = 0x8b;
+    private static final Byte opcode = (byte) 0x8b;
 
     public Op1Add() {
         super(word, opcode);
