+++ src/com/timmattison/bitcoin/test/script/words/arithmetic/Op2Div.java
@@ -12,7 +12,7 @@ import com.timmattison.bitcoin.test.script.Word;
  */
 public class Op2Div extends Word {
     private static final String word = "OP_2DIV";
-    private static final int opcode = 0x8e;
+    private static final Byte opcode = (byte) 0x8e;
 
     public Op2Div() {
         super(word, opcode);
