+++ src/com/timmattison/bitcoin/test/script/words/arithmetic/Op2Mul.java
@@ -12,7 +12,7 @@ import com.timmattison.bitcoin.test.script.Word;
  */
 public class Op2Mul extends Word {
     private static final String word = "OP_2MUL";
-    private static final int opcode = 0x8d;
+    private static final Byte opcode = (byte) 0x8d;
 
     public Op2Mul() {
         super(word, opcode);
