+++ src/com/timmattison/bitcoin/test/script/words/arithmetic/OpAdd.java
@@ -12,7 +12,7 @@ import com.timmattison.bitcoin.test.script.Word;
  */
 public class OpAdd extends Word {
     private static final String word = "OP_ADD";
-    private static final int opcode = 0x93;
+    private static final Byte opcode = (byte) 0x93;
 
     public OpAdd() {
         super(word, opcode);
