+++ src/com/timmattison/bitcoin/test/script/words/arithmetic/OpNegate.java
@@ -12,7 +12,7 @@ import com.timmattison.bitcoin.test.script.Word;
  */
 public class OpNegate extends Word {
     private static final String word = "OP_NEGATE";
-    private static final int opcode = 0x8f;
+    private static final Byte opcode = (byte) 0x8f;
 
     public OpNegate() {
         super(word, opcode);
