+++ src/com/timmattison/bitcoin/test/script/words/arithmetic/OpNot.java
@@ -12,7 +12,7 @@ import com.timmattison.bitcoin.test.script.Word;
  */
 public class OpNot extends Word {
     private static final String word = "OP_NOT";
-    private static final int opcode = 0x91;
+    private static final Byte opcode = (byte) 0x91;
 
     public OpNot() {
         super(word, opcode);
