+++ src/com/timmattison/bitcoin/test/script/words/arithmetic/OpNumEqual.java
@@ -12,7 +12,7 @@ import com.timmattison.bitcoin.test.script.Word;
  */
 public class OpNumEqual extends Word {
     private static final String word = "OP_NUMEQUAL";
-    private static final int opcode = 0x9c;
+    private static final Byte opcode = (byte) 0x9c;
 
     public OpNumEqual() {
         super(word, opcode);
