+++ src/com/timmattison/bitcoin/test/script/words/arithmetic/OpNumNotEqual.java
@@ -12,7 +12,7 @@ import com.timmattison.bitcoin.test.script.Word;
  */
 public class OpNumNotEqual extends Word {
     private static final String word = "OP_NUMNOTEQUAL";
-    private static final int opcode = 0x9e;
+    private static final Byte opcode = (byte) 0x9e;
 
     public OpNumNotEqual() {
         super(word, opcode);
