+++ src/com/timmattison/bitcoin/test/script/words/bitwiselogic/OpEqualVerify.java
@@ -12,7 +12,7 @@ import com.timmattison.bitcoin.test.script.Word;
  */
 public class OpEqualVerify extends Word {
     private static final String word = "OP_EQUALVERIFY";
-    private static final int opcode = 0x88;
+    private static final Byte opcode = (byte) 0x88;
 
     public OpEqualVerify() {
         super(word, opcode);
