+++ src/com/timmattison/bitcoin/test/script/words/bitwiselogic/OpInvert.java
@@ -12,7 +12,7 @@ import com.timmattison.bitcoin.test.script.Word;
  */
 public class OpInvert extends Word {
     private static final String word = "OP_INVERT";
-    private static final int opcode = 0x83;
+    private static final Byte opcode = (byte) 0x83;
 
     public OpInvert() {
         super(word, opcode);
