+++ src/com/timmattison/bitcoin/test/script/words/crypto/OpCheckMultiSigVerify.java
@@ -12,7 +12,7 @@ import com.timmattison.bitcoin.test.script.Word;
  */
 public class OpCheckMultiSigVerify extends Word {
     private static final String word = "OP_CHECKMULTISIGVERIFY";
-    private static final int opcode = 0xaf;
+    private static final Byte opcode = (byte) 0xaf;
 
     public OpCheckMultiSigVerify() {
         super(word, opcode);
