+++ src/com/timmattison/bitcoin/test/script/words/crypto/OpHash160.java
@@ -12,7 +12,7 @@ import com.timmattison.bitcoin.test.script.Word;
  */
 public class OpHash160 extends Word {
     private static final String word = "OP_HASH160";
-    private static final int opcode = 0xa9;
+    private static final Byte opcode = (byte) 0xa9;
 
     public OpHash160() {
         super(word, opcode);
