+++ src/com/timmattison/bitcoin/test/script/words/flowcontrol/OpElse.java
@@ -12,7 +12,7 @@ import com.timmattison.bitcoin.test.script.Word;
  */
 public class OpElse extends Word {
     private static final String word = "OP_ELSE";
-    private static final int opcode = 0x67;
+    private static final Byte opcode = (byte) 0x67;
 
     public OpElse() {
         super(word, opcode);
