+++ src/com/timmattison/bitcoin/test/script/words/flowcontrol/OpNop.java
@@ -12,7 +12,7 @@ import com.timmattison.bitcoin.test.script.Word;
  */
 public class OpNop extends Word {
     private static final String word = "OP_NOP";
-    private static final int opcode = 0x61;
+    private static final Byte opcode = (byte) 0x61;
 
     public OpNop() {
         super(word, opcode);
