+++ src/com/timmattison/bitcoin/test/script/words/flowcontrol/OpNotIf.java
@@ -12,7 +12,7 @@ import com.timmattison.bitcoin.test.script.Word;
  */
 public class OpNotIf extends Word {
     private static final String word = "OP_NOTIF";
-    private static final int opcode = 0x64;
+    private static final Byte opcode = (byte) 0x64;
 
     public OpNotIf() {
         super(word, opcode);
