+++ src/com/timmattison/bitcoin/test/script/words/flowcontrol/OpVerify.java
@@ -12,7 +12,7 @@ import com.timmattison.bitcoin.test.script.Word;
  */
 public class OpVerify extends Word {
     private static final String word = "OP_VERIFY";
-    private static final int opcode = 0x69;
+    private static final Byte opcode = (byte) 0x69;
 
     public OpVerify() {
         super(word, opcode);
