+++ src/com/timmattison/bitcoin/test/script/words/reservedwords/OpNop6.java
@@ -12,7 +12,7 @@ import com.timmattison.bitcoin.test.script.Word;
  */
 public class OpNop6 extends Word {
     private static final String word = "OP_NOP6";
-    private static final int opcode = 0xb5;
+    private static final Byte opcode = (byte) 0xb5;
 
     public OpNop6() {
         super(word, opcode);
