+++ src/com/timmattison/bitcoin/test/script/words/reservedwords/OpNop7.java
@@ -12,7 +12,7 @@ import com.timmattison.bitcoin.test.script.Word;
  */
 public class OpNop7 extends Word {
     private static final String word = "OP_NOP7";
-    private static final int opcode = 0xb6;
+    private static final Byte opcode = (byte) 0xb6;
 
     public OpNop7() {
         super(word, opcode);
