+++ src/com/timmattison/bitcoin/test/script/words/reservedwords/OpNop9.java
@@ -12,7 +12,7 @@ import com.timmattison.bitcoin.test.script.Word;
  */
 public class OpNop9 extends Word {
     private static final String word = "OP_NOP9";
-    private static final int opcode = 0xb8;
+    private static final Byte opcode = (byte) 0xb8;
 
     public OpNop9() {
         super(word, opcode);
