+++ src/com/timmattison/bitcoin/test/script/words/reservedwords/OpReserved1.java
@@ -12,7 +12,7 @@ import com.timmattison.bitcoin.test.script.Word;
  */
 public class OpReserved1 extends Word {
     private static final String word = "OP_RESERVED1";
-    private static final int opcode = 0x89;
+    private static final Byte opcode = (byte) 0x89;
 
     public OpReserved1() {
         super(word, opcode);
