+++ src/com/timmattison/bitcoin/test/script/words/splice/OpCat.java
@@ -12,7 +12,7 @@ import com.timmattison.bitcoin.test.script.Word;
  */
 public class OpCat extends Word {
     private static final String word = "OP_CAT";
-    private static final int opcode = 0x7e;
+    private static final Byte opcode = (byte) 0x7e;
 
     public OpCat() {
         super(word, opcode);
