+++ src/com/timmattison/bitcoin/test/script/words/stack/Op2Drop.java
@@ -12,7 +12,7 @@ import com.timmattison.bitcoin.test.script.Word;
  */
 public class Op2Drop extends Word {
     private static final String word = "OP_2DROP";
-    private static final int opcode = 0x6d;
+    private static final Byte opcode = (byte) 0x6d;
 
     public Op2Drop() {
         super(word, opcode);
