+++ src/com/timmattison/bitcoin/test/script/words/stack/Op2Dup.java
@@ -12,7 +12,7 @@ import com.timmattison.bitcoin.test.script.Word;
  */
 public class Op2Dup extends Word {
     private static final String word = "OP_2DUP";
-    private static final int opcode = 0x6e;
+    private static final Byte opcode = (byte) 0x6e;
 
     public Op2Dup() {
         super(word, opcode);
